Update Readme.md
## Background
I had been using Jackalope for Fuzzing and to confirm that it could find easy to identify Bugs.  Looking deeper I found AUF, OOB, NPTR and other issues that concealed Bugs. 

This Project is for anyone wanting to Learn & Educate on XNU Image Fuzzing. You can use Jackalope with the Example code provided in Imageio, and have your own Stand Alone XNU Image Fuzzer to expand your horizons.

Are you new to XNU? 

This Code is for you!

If you have Questions Open an Issue. This Code is meant to stimulate discussion on the Bugs that can be hit from the Remote, No-Auth position.